javadoc for RetryReason; package name change
