+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/retryReasonCategories/ConnectionResetRetryReason.java
@@ -16,29 +16,23 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.fs.azurebfs.services.retryReasonAbbreviationCreation;
+package org.apache.hadoop.fs.azurebfs.services.retryReasonCategories;
 
-import java.io.IOException;
+import static org.apache.hadoop.fs.azurebfs.services.RetryReasonConstants.connectionResetMessage;
 
-import org.apache.hadoop.fs.azurebfs.services.retryReasonAbbreviationCreation.RetryReasonAbbreviationCreator;
-
-
-public class UnknownIOExceptionRetryReason extends
-    RetryReasonAbbreviationCreator {
+public class ConnectionResetRetryReason extends
+    RetryReasonCategory {
 
   @Override
   public Boolean canCapture(final Exception ex,
       final Integer statusCode,
       final String serverErrorMessage) {
-    if (ex instanceof IOException) {
-      return true;
-    }
-    return false;
+    return checkExceptionMessage(ex, connectionResetMessage);
   }
 
   @Override
   public String getAbbreviation(final Integer statusCode,
       final String serverErrorMessage) {
-    return "IOE";
+    return "CR";
   }
 }
